+++ src/main/java/com/apr/car_sales/persistence/bid/BidRepository.java
@@ -2,5 +2,5 @@ package com.apr.car_sales.persistence.bid;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BidRepository extends JpaRepository<BidEntity, Integer> {
+public interface BidRepository extends JpaRepository<BidEntity, Long> {
 }
