+++ src/main/java/com/apr/car_sales/persistence/car/CarRepository.java
@@ -2,6 +2,6 @@ package com.apr.car_sales.persistence.car;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<CarEntity, Integer> {
+public interface CarRepository extends JpaRepository<CarEntity, Long> {
 
 }
