+++ src/main/java/com/apr/car_sales/persistence/purchase/PaymentEntity.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 @Data
 @Entity
@@ -13,7 +12,7 @@ import java.time.format.DateTimeFormatter;
 public class PaymentEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private double amount;
     private LocalDateTime paymentDate;
